fix status for invoice updation

+++ routes/api.php
@@ -13,18 +13,25 @@ Route::get('/user', function (Request $request) {
 
 
 //INVOICE
-Route::post('/invoice',[InvoiceController::class, 'store'])->middleware('role:admin,accountant,user');
-Route::get('/invoices',[InvoiceController::class, 'index'])->middleware('role:admin,accountant');
+Route::post('/invoice',[InvoiceController::class, 'store'])->middleware(['auth:sanctum','role:admin,accountant,user']);
+Route::get('/invoices',[InvoiceController::class, 'index'])->middleware(['auth:sanctum','role:admin,accountant']);
+Route::put('/invoice/{id}',[InvoiceController::class, 'update'])->middleware(['auth:sanctum','role:admin,accountant']);
+Route::delete('/invoice/{id}',[InvoiceController::class, 'delete'])->middleware(['auth:sanctum','role:admin,accountant']);
+
+// Route::post('/invoice',[InvoiceController::class, 'store']);
+// Route::get('/invoices',[InvoiceController::class, 'index']);
+// Route::put('/invoice/{id}',[InvoiceController::class, 'update']);
+// Route::delete('/invoice/{id}',[InvoiceController::class, 'delete']);
 
 //PAYMENT
-Route::post('/payment',[PaymentController::class, 'store'])->middleware('role:admin,accountant,user');
+Route::post('/payment',[PaymentController::class, 'store'])->middleware(['auth:sanctum','role:admin,accountant,user']);
 
 //AUTHENTICATION 
-Route::post('/register',[UserController::class, 'register'])->middleware('role:admin,accountant,user');
-Route::post('/login',[UserController::class, 'login'])->middleware('role:admin,accountant,user');
-Route::post('/logout',[UserController::class, 'logout'])->middleware('role:admin,accountant,user');
+Route::post('/register',[UserController::class, 'register']);
+Route::post('/login',[UserController::class, 'login']);
+Route::post('/logout',[UserController::class, 'logout']);
 
 //USER
-Route::get('/users',[UserController::class, 'index'])->middleware('role:admin,accountant');
-Route::put('/user/{id}',[UserController::class, 'update']) ->middleware('role:admin');
-Route::delete('/user/{id}',[UserController::class, 'destroy']) ->middleware('role:admin');
+Route::get('/users',[UserController::class, 'index'])->middleware(['auth:sanctum','role:admin,accountant']);
+Route::put('/user/{id}',[UserController::class, 'update']) ->middleware(['auth:sanctum','role:admin']);
+Route::delete('/user/{id}',[UserController::class, 'destroy']) ->middleware(['auth:sanctum','role:admin']);
